domain 패키지와 view 패키지를 생성하고 해당하는 클래스를 옮겨줬다.
domain 패키지(핵심 비지니스 로직) : Controller, Ladder, LadderRow, Names
view 패키지(UI 관련 클래스) : Input
