+++ src/main/java/kr/codesquad/Main.java
@@ -1,5 +1,7 @@
 package kr.codesquad;
 
+import kr.codesquad.domain.Controller;
+
 public class Main {
     public static void main(String[] args) {
         Controller controller = new Controller();
