+++ src/main/java/kr/codesquad/domain/Controller.java
@@ -1,4 +1,7 @@
-package kr.codesquad;
+package kr.codesquad.domain;
+
+import kr.codesquad.domain.Ladder;
+import kr.codesquad.view.Input;
 
 import java.io.IOException;
